[restructure] Change `yamlProjectConfiguration` to `projectConfiguration/yaml`
This is needed to prepare for the new project configuration `genezio.config.ts` type.
