+++ src/generateSdk/sdkGenerator/TsSdkGenerator.ts
@@ -22,7 +22,7 @@ import {
     MapType,
     SdkGeneratorClassesInfoInput,
 } from "../../models/genezioModels.js";
-import { TriggerType } from "../../yamlProjectConfiguration/models.js";
+import { TriggerType } from "../../projectConfiguration/yaml/models.js";
 import { storageTs } from "../templates/nodeSdkTs.js";
 import path from "path";
