+++ src/generateSdk/utils/getAstSummary.ts
@@ -6,7 +6,7 @@ import {
     ParameterDefinition,
     SdkGeneratorClassesInfoInput,
 } from "../../models/genezioModels.js";
-import { TriggerType } from "../../yamlProjectConfiguration/models.js";
+import { TriggerType } from "../../projectConfiguration/yaml/models.js";
 
 export function getAstSummary(classesInfo: SdkGeneratorClassesInfoInput[]): AstSummaryClass[] {
     const classes: AstSummaryClass[] = classesInfo
